optionally replace the filtering stage by a neural network (bb_wdd_filter)
The neural network should be faster and more accurate compared to the old manual stage.
However, the network currently does not give a reliable estimate for the waggle duration as it has a too narrow FoV.

The change also comes with the feature to only save (/and transmit over IPC) the actual waggles and disregard all other detections. And, for evaluation purposes, it's also possible to save nothing at all.